Added Rhythm Generator for Full Circle
This was going to be used for our audio processing library. However, we
haven't used it. From this point on this repository is going to be used
as the full circle rhythm generator repository. This commit adds the
basic layout of the rhythm generator with a working song (Do I Wanna
Know) and a working mechanic for generating rhythms. This may or may not
be updated depending on how often it is used and whether our gameplay
mechanics change in the near future.